requirements.txt
# Auto-Debug Agent — Python dependencies
# All other imports are Python standard library.
#
# Install:
#   pip install -r requirements.txt

anthropic>=0.76.0
python-dotenv>=1.0.1
